requirements.txt
PySide6>=6.6
Pillow>=9.0
google-genai>=0.4.0
numpy
